chave estrangeira feita

+++ backend/api/src/controllers/sala-controller/delete-sala-controller.js
@@ -1,5 +1,7 @@
 const { HttpHelper } = require('../../utils/http-helper');
 const { SalaModel } = require('../../models/sala-model');
+const { ReservaModel } = require('../../models/reserva-model')
+const { Op } = require("sequelize");
 
 class DeleteSalaController {
        
@@ -7,10 +9,21 @@ class DeleteSalaController {
         const httpHelper = new HttpHelper(response);
         try {
             const { id } = request.params;
+
+            const restricao = await ReservaModel.findAll({where: {
+                SalaId:{
+                    [Op.eq]: id
+                }
+            }})
+
+            if(restricao[0]) return httpHelper.badRequest("Solicitação negada. Ainda há reservas utilizando esta sala")
+
             if (!id) return httpHelper.badRequest('Parâmetros inválidos!');
             const salaExists = await SalaModel.findOne({ where: { id } });
             if (!salaExists) return httpHelper.notFound('Sala não encontrada!');
+            
             await SalaModel.destroy({ where: { id } });
+
             return httpHelper.ok({
                 message: 'Sala deletada com sucesso!'
             })
